[PR80025] avoid cselib rtx_equal infinite recursion on XOR
When two VALUEs are recorded in the cselib equivalence table such that
they are equivalent to each other XORed with the same expression, if
we started a cselib equivalence test between say the odd one and the
even one, we'd end up recursing to compare the even one with the odd
one, and then again, indefinitely.

This patch cuts off this infinite recursion by recognizing the XOR
special case (it's the only binary commutative operation in which
applying one of the operands of an operation to the result of that
operation brings you back to the other operand) and determining
whether we have an equivalence without recursing down the infinite
path.

for  gcc/ChangeLog

	* cselib.c (rtx_equal_for_cselib_1): Avoid infinite recursion
        on XOR.

for  gcc/testsuite/ChangeLog

	* gcc.dg/torture/pr80025.c: New.